requirements-desktop.txt
-r requirements.txt
waitress
pywebview
pyinstaller
